+++ config/config.php
@@ -5,10 +5,13 @@ use ProtoneMedia\LaravelFormComponents\Components;
 return [
     'prefix' => '',
 
-    /** tailwind | tailwind-2 | bootstrap-4 | bootstrap-5 */
+    /** tailwind | tailwind-2 | tailwind-forms-simple | bootstrap-4 | bootstrap-5 */
     'framework' => 'tailwind',
 
     'use_eloquent_date_casting' => false,
+
+    /** bool | string */
+    'default_wire' => false,
 
     'components' => [
         'form' => [