+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `laravel-form-components` will be documented in this file
 
+## 3.5.0 - 2022-01-05
+
+- Added `default_wire` configuration
+- Fix for overriding `wire:model` attribute
+
+## 3.4.0 - 2022-01-04
+
+- Added `tailwind-forms-simple` views
+
 ## 3.3.1 - 2021-12-22
 
 - Ensure nested properties uses dot notation